+++ src/components/ui/common/Buttons.tsx
@@ -1,6 +1,7 @@
 'use client';
 import React from 'react';
 import { motion } from 'framer-motion';
+
 import { cn } from '@/lib/utils';
 
 export interface ButtonsProps {
